mypage 요청 받는거

+++ mypage-service/src/main/java/com/traplaner/mypageservice/mypage/client/TravelPlanServiceClient.java
@@ -19,7 +19,7 @@ public interface TravelPlanServiceClient {
     CommonResDto<List<travelPlanResDto>> findByMemberId(@PathVariable Integer id);
 
     @GetMapping("journeysByTravelId/{travelId}")
-    travelPlanResDto findById(@PathVariable Long travelId);
+    travelPlanResDto findById(@PathVariable int travelId);
 
     @GetMapping("/travelListsByMemberId?memberId={memberId}&page={page}&size={size}")
     CommonResDto<Page<travelPlanResDto>> findByMemberId(@PathVariable Integer memberId, @PathVariable int page, @PathVariable int size);
